Check for <time.h> and <sys/time.h> in configure. Remove AC_HEADER_TIME (no longer necessary except for the very oldest operating systems to the Autoconf manual).
Regenerate configure and config.h.in.

Rename sys directory to 'missing' to be a place to keep missing header
files, etc. Calling it sys confuses cpp when #including standard sys/
headers.
