+++ nlopt/config.h.in
@@ -54,11 +54,17 @@
 /* Define to 1 if you have the <sys/stat.h> header file. */
 #undef HAVE_SYS_STAT_H
 
+/* Define to 1 if you have the <sys/time.h> header file. */
+#undef HAVE_SYS_TIME_H
+
 /* Define to 1 if you have the <sys/types.h> header file. */
 #undef HAVE_SYS_TYPES_H
 
 /* Define to 1 if you have the `time' function. */
 #undef HAVE_TIME
+
+/* Define to 1 if you have the <time.h> header file. */
+#undef HAVE_TIME_H
 
 /* Define to 1 if the system has the type `uint32_t'. */
 #undef HAVE_UINT32_T
@@ -96,9 +102,6 @@
 /* Define to C thread-local keyword, or to nothing if this is not supported in
    your compiler. */
 #undef THREADLOCAL
-
-/* Define to 1 if you can safely include both <sys/time.h> and <time.h>. */
-#undef TIME_WITH_SYS_TIME
 
 /* Define to empty if `const' does not conform to ANSI C. */
 #undef const
