+++ nlopt/missing/time.h
@@ -1,0 +1,7 @@
+#pragma once
+ 
+#ifndef _TIMES_H
+ 
+#include "sys/times.h"
+ 
+#endif
