--- nlopt/sys/time.cpp
@@ -1,18 +1,0 @@
-#include "sys/times.h"
- 
-int gettimeofday(struct timeval* t,void*)
-{       struct _timeb timebuffer;
-        _ftime( &timebuffer );
-        t->tv_sec=(long)timebuffer.time;
-        t->tv_usec=1000*timebuffer.millitm;
-		return 0;
-}
- 
-clock_t times (struct tms *__buffer) {
- 
-	__buffer->tms_utime = clock();
-	__buffer->tms_stime = 0;
-	__buffer->tms_cstime = 0;
-	__buffer->tms_cutime = 0;
-	return __buffer->tms_utime;
-}
