--- nlopt/sys/time.h
@@ -1,7 +1,0 @@
-#pragma once
- 
-#ifndef _TIMES_H
- 
-#include "sys/times.h"
- 
-#endif
